replaced images, updated styles
